split AST declaration support types
